#11 Move Login and CreateUser to UserAccess folder
In order re-organize the components
1. Move the LoginComponent and CreateUserComponent to a new subfolder 'UserAccess'
