+++ fmt-ui/src/app/app.module.ts
@@ -4,11 +4,11 @@ import {HttpClientModule} from "@angular/common/http";
 
 import {AppRoutingModule} from './app-routing.module';
 import {AppComponent} from './app.component';
-import {LoginComponent} from './components/login/login.component';
+import {LoginComponent} from './components/user-access/login/login.component';
 import {HomeComponent} from './components/home/home.component';
 import {FormsModule, ReactiveFormsModule} from "@angular/forms";
 import {CommonModule} from "@angular/common";
-import {CreateUserComponent} from "./components/create-user/create-user.component";
+import {CreateUserComponent} from "./components/user-access/create-user/create-user.component";
 import {UserHomePage} from './components/feedback/user-home-page/user-home-page.component';
 import {DataTransferService} from "./services/data-transfer.service";
