+++ fmt-ui/src/app/components/user-access/create-user/create-user.component.ts
@@ -1,9 +1,9 @@
 import {Component, OnInit} from '@angular/core';
-import {InputValidator} from "../../helpers/input-validator";
-import {CreateUserService} from "../../services/create-user.service";
-import {User} from "../../domain/user";
+import {InputValidator} from "../../../helpers/input-validator";
+import {CreateUserService} from "../../../services/create-user.service";
+import {User} from "../../../domain/user";
 import {Router} from "@angular/router";
-import {DataTransferService} from "../../services/data-transfer.service";
+import {DataTransferService} from "../../../services/data-transfer.service";
 
 @Component({
   selector: 'app-create-user',
